SEDRFuse-main/requirements.txt
Ubuntu 16.04
CUDA 9.0.176
CUDNN 7.0.5
python=3.6.2


tensorboard==1.7.0
tensorflow-gpu==1.7.0
opencv-python==4.1.2.30
numpy==1.13.1
matplotlib==3.1.2
pandas==0.24.2
Pillow==6.2.1
scikit-image==0.16.2
scipy==1.2.1








